+++ vitrina/structure/templates/vitrina/structure/dataset_structure.html
@@ -192,13 +192,15 @@
                         {% endif %}
                     </div>
                     <div class="column is-1">
+                    {% if metadata.metadata_version %}
                         {% if metadata.metadata_version.status == "DRAFT" %}
                             <span class="tag is-danger is-light">({% translate "rengiama" %})</span>
-                        {% elif metadata.metadata_version and metadata.draft and metadata.metadata_version.status != "DRAFT"%}
+                        {% elif metadata.draft and metadata.metadata_version.status != "DRAFT"%}
                             <span class="tag is-danger is-light">{{ metadata.metadata_version.external_version}} ({% translate "rengiama" %})</span>
                         {% elif metadata.metadata_version.external_version and not metadata.draft %}
                             <span class="tag is-success is-light">{{ metadata.metadata_version.external_version }}</span>
                         {% endif %}
+                    {% endif %}
                     </div>
                 </div>
         {% endfor %}
