+++ vitrina/structure/templates/vitrina/structure/model_structure.html
@@ -42,13 +42,16 @@
             {% endif %}
             <span>{{ metadata.average_level|default:"0"|floatformat:"0" }}</span>
             <span class="icon has-text-warning pr-3"><i class="fas fa-star"></i></span>
+            {% if metadata.metadata_version %}
 
-            {% if metadata.metadata_version.status == "DRAFT" %}
-                <span class="tag is-danger is-light">({% translate "rengiama" %})</span>
-            {% elif metadata.metadata_version and metadata.draft and metadata.metadata_version.status != "DRAFT"%}
-                <span class="tag is-danger is-light">{{ metadata.metadata_version.external_version}} ({% translate "rengiama" %})</span>
-            {% elif metadata.metadata_version.external_version and not metadata.draft %}
-                <span class="tag is-success is-light">{{ metadata.metadata_version.external_version }}</span>
+                {% if metadata.metadata_version.status == "DRAFT" %}
+                    <span class="tag is-danger is-light">({% translate "rengiama" %})</span>
+                {% elif metadata.draft and metadata.metadata_version.status != "DRAFT"%}
+                    <span class="tag is-danger is-light">{{ metadata.metadata_version.external_version}} ({% translate "rengiama" %})</span>
+                {% elif metadata.metadata_version.external_version and not metadata.draft %}
+                    <span class="tag is-success is-light">{{ metadata.metadata_version.external_version }}</span>
+                {% endif %}
+
             {% endif %}
 
             {% if metadata.status.codename %}
@@ -294,19 +297,18 @@
                         {% endif %}
                     </div>
                         <div class="column is-1">
+                        {% if prop_metadata.metadata_version %}
+
                             {% if prop_metadata.metadata_version.status == "DRAFT" %}
                                 <span class="tag is-danger is-light">({% translate "rengiama" %})</span>
-                            {% elif prop_metadata.metadata_version and prop_metadata.draft and prop_metadata.metadata_version.status != "DRAFT"%}
+                            {% elif prop_metadata.draft and prop_metadata.metadata_version.status != "DRAFT"%}
                                 <span class="tag is-danger is-light">{{ prop_metadata.metadata_version.external_version}} ({% translate "rengiama" %})</span>
                             {% elif prop_metadata.metadata_version.external_version and not prop_metadata.draft %}
                                 <span class="tag is-success is-light">{{ prop_metadata.metadata_version.external_version }}</span>
 
                             {% endif %}
-
-
-
+                        {% endif %}
                         </div>
-
                     </div>
                 </div>
                 {% endfor %}
@@ -409,13 +411,17 @@
                                 </div>
                                 {% if prop.model == model %}
                                 <div class="column is-1">
+                                    {% if prop_metadata.metadata_version %}
+
                                         {% if prop_metadata.metadata_version.status == "DRAFT" %}
                                             <span class="tag is-danger is-light">({% translate "rengiama" %})</span>
-                                        {% elif prop_metadata.metadata_version and prop_metadata.draft and prop_metadata.metadata_version.status != "DRAFT"%}
+                                        {% elif prop_metadata.draft and prop_metadata.metadata_version.status != "DRAFT"%}
                                             <span class="tag is-danger is-light">{{ prop_metadata.metadata_version.external_version}} ({% translate "rengiama" %})</span>
                                         {% elif prop_metadata.metadata_version.external_version and not prop_metadata.draft %}
                                             <span class="tag is-success is-light">{{ prop_metadata.metadata_version.external_version }}</span>
                                         {% endif %}
+
+                                    {% endif %}
                                 </div>
                                 {% endif %}
                                 {% if prop.model == model %}
