+++ vitrina/structure/templates/vitrina/structure/property_structure.html
@@ -84,12 +84,14 @@
             <span class="icon has-text-warning"><i class="fas fa-star"></i></span>
             {% endif %}
 
-            {% if metadata.metadata_version.status == "DRAFT" %}
-                <span class="tag is-danger is-light">({% translate "rengiama" %})</span>
-            {% elif metadata.metadata_version and metadata.draft and metadata.metadata_version.status != "DRAFT"%}
-                <span class="tag is-danger is-light">{{ metadata.metadata_version.external_version}} ({% translate "rengiama" %})</span>
-            {% elif metadata.metadata_version.external_version and not metadata.draft %}
-                <span class="tag is-success is-light">{{ metadata.metadata_version.external_version }}</span>
+            {% if metadata.metadata_version %}
+                {% if metadata.metadata_version.status == "DRAFT" %}
+                    <span class="tag is-danger is-light">({% translate "rengiama" %})</span>
+                {% elif metadata.draft and metadata.metadata_version.status != "DRAFT"%}
+                    <span class="tag is-danger is-light">{{ metadata.metadata_version.external_version}} ({% translate "rengiama" %})</span>
+                {% elif metadata.metadata_version.external_version and not metadata.draft %}
+                    <span class="tag is-success is-light">{{ metadata.metadata_version.external_version }}</span>
+                {% endif %}
             {% endif %}
 
             {% if metadata.status.codename %}
@@ -288,13 +290,19 @@
                             {% endif %}
                         </div>
                         <div class="column is-1">
+
+                        {% if enum_meta.metadata_version %}
+
                             {% if enum_meta.metadata_version.status == "DRAFT" %}
                                 <span class="tag is-danger is-light">({% translate "rengiama" %})</span>
-                            {% elif enum_meta.metadata_version and enum_meta.draft and enum_meta.metadata_version.status != "DRAFT"%}
+                            {% elif enum_meta.draft and enum_meta.metadata_version.status != "DRAFT"%}
                                 <span class="tag is-danger is-light">{{ enum_meta.metadata_version.external_version}} ({% translate "rengiama" %})</span>
                             {% elif enum_meta.metadata_version.external_version and not enum_meta.draft %}
                                 <span class="tag is-success is-light">{{ enum_meta.metadata_version.external_version }}</span>
                             {% endif %}
+
+                        {% endif %}
+
                         </div>
                         <div class="column is-2">
                             {% if enum_meta.prepare or enum_meta.title %}
